requirements.txt
fastapi
uvicorn[standard]
langchain
sentence-transformers
openai
requests
python-dotenv
django
aiohttp
python-multipart
tqdm
PyPDF2
transformers
sentencepiece
chromadb==0.5.3
